refactor: add signup with google logic to develop
