+++ apps/freelance/src/app/app.tsx
@@ -3,13 +3,13 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import LoginPage from '@pages/login-page/login-page';
 import SignupPage from '@pages/signup-page/signup-page';
-import store from '../redux/example-store';
+import store from '../redux/store';
 import { GoogleOAuthProvider } from '@react-oauth/google';
 
 export function App() {
   return (
     <StyledApp>
-      <GoogleOAuthProvider clientId ="253619542281-miag0ub2aorap933406d1vg3bbte7j8m.apps.googleusercontent.com"> 
+      <GoogleOAuthProvider clientId={`${process.env['NX_APP_GOOGLE_KEY']}`}> 
        <Provider store={store}>
         <BrowserRouter>
         <Routes>
